Update Parsha Sheet Emailer.gs
Added code for Rosh Hashanah, Sukkos, and Chanukah

+++ Parsha Sheet Emailer.gs
@@ -1,16 +1,18 @@
-function main() {  
-  var d = new Date();
-  if (d.getFullYear() != PropertiesService.getScriptProperties().getProperty("thisYear")) {
-  PropertiesService.getScriptProperties().deleteAllProperties();
-  PropertiesService.getScriptProperties().setProperty("thisYear", d.getFullYear());
+function main() {
+  var d = new Date();  
+  var hebrewDate = getHebrewDate(d);
+  const hebrewMonth = hebrewDate.get("Month");
+  const hebrewDay = hebrewDate.get("Day");
+  const hebrewYear = hebrewDate.get("Year");
+  if (hebrewMonth == "Cheshvan") {
+    PropertiesService.getScriptProperties().deleteAllProperties();
   }
-  var year = d.getFullYear();  
   var parsha = [];
   var holiday = "";
   var hebcal = getHebcal(d);
   var thisWeeksParsha = hebcal.items.find(item => item.title.includes("Parshas "));
   if (thisWeeksParsha) {
-    var thisWeeksParsha = thisWeeksParsha.title.replaceAll("Parshas ", "").replaceAll("Sazria","Tazria")
+    var thisWeeksParsha = thisWeeksParsha.title.replaceAll("Parshas ", "").replaceAll("Sazria", "Tazria")
     if (thisWeeksParsha.includes("-") && thisWeeksParsha != "Lech-Lecha") {
       parsha = thisWeeksParsha.split("-");
     }
@@ -19,15 +21,12 @@ function main() {
     }
   }
   if (hebcal.items.find(item => item.title.includes("Simchas Torah"))) {
-    if (PropertiesService.getScriptProperties().getProperty(`V'Zos Habracha${year}`) != "sent") {
-      PropertiesService.getScriptProperties().setProperty(`V'Zos Habracha${year}`, "sent");
+    if (PropertiesService.getScriptProperties().getProperty(`V'Zos Habracha${hebrewYear}`) != "sent") {
+      PropertiesService.getScriptProperties().setProperty(`V'Zos Habracha${hebrewYear}`, "sent");
       parsha.push("V'Zos Habracha");
     }
-  }  
-  var holiday = hebcal.items.find(item => item.category.includes("holiday"));
-  if (holiday) {    
-    holiday = getHoliday(holiday.title, year);
   }
+  holiday = getHoliday(hebrewMonth, hebrewDay, hebrewYear);
   if (parsha.length != 0 || holiday != "") {
     emailSheet(parsha, holiday);
   }
@@ -36,21 +35,55 @@ function main() {
   }
 }
 
-function getHoliday(title, year) {  
-  title = title.replaceAll("Erev ", "");
-  if (title === "Shavuos" || title === "Pesach" || title === "Purim") {
-    var property = title+year;
-    PropertiesService.getScriptProperties().getProperty(property) != "sent" ?
-      PropertiesService.getScriptProperties().setProperty(property, "sent") : title = "";   
+function getHoliday(month, day, year) {
+  var holiday = "";
+  month = String(month);
+  if (month == "Elul" && day >= 20) {
+    holiday = "Rosh Hashanah/Yom Kippur";
   }
-  else {
-    title = "";
+  else if (month == "Tishrei" && day >= 5) {
+    holiday = "Sukkos";
   }
-  return title;
+  else if (month == "Kislev" && day >= 15) {
+    holiday = "Chanukah";
+  }
+  else if ((month.startsWith("Adar") && month != 'Adar I') && day >= 4) {
+    holiday = "Megillah"
+  }
+  else if (month == "Nissan" && day >= 5) {
+    holiday = "Haggadah";
+  }
+  else if ((month == "Iyar" && day >= 26) || (month == "Sivan" && day <= 6)) {
+    holiday = "Rus";
+  }
+
+  if (holiday != "") {
+    var property = holiday + year;
+    if (PropertiesService.getScriptProperties().getProperty(property) != "sent") {
+      PropertiesService.getScriptProperties().setProperty(property, "sent");
+    }
+    else {
+      holiday = "";
+    }
+  }
+  return holiday;
 }
 
 function getHebcal(todaysDate) {
   return JSON.parse(UrlFetchApp.fetch(`https://www.hebcal.com/hebcal?v=1&cfg=json&maj=on&min=on&mod=off&nx=off&start=${getDate(todaysDate)}&end=${getDate(addDays(todaysDate, 10))}&ss=off&lg=a&mf=off&c=off&M=off&s=on`).getContentText());
+}
+
+function getHebrewDate(d) {
+  var month = (d.getMonth() + 1).toString().padStart(2, '0');
+  var day = d.getDate().toString().padStart(2, '0');
+  var hebcal = JSON.parse(UrlFetchApp.fetch(`https://www.hebcal.com/converter?cfg=json&gy=${d.getFullYear()}&gm=${month}&gd=${day}&g2h=1`).getContentText());
+  var hebrewMonth = hebcal.hm;
+  hebrewMonth = hebrewMonth.replaceAll("Nisan", "Nissan").replaceAll("Iyyar", "Iyar").replaceAll("Tevet", "Teves").replaceAll("Sh'vat", "Shvat");
+  const hebrewDate = new Map();
+  hebrewDate.set("Month", hebrewMonth);
+  hebrewDate.set("Day", hebcal.hd);
+  hebrewDate.set("Year", hebcal.hy);
+  return hebrewDate;
 }
 
 const getDate = d => `${d.getFullYear()}-${(d.getMonth() + 1).toString().padStart(2, '0')}-${d.getDate().toString().padStart(2, '0')}`;
@@ -66,35 +99,22 @@ function emailSheet(parsha, holiday) {
   if (parsha.length == 2) {
     subject = `Parsha Questions - ${parsha[0]}/${parsha[1]}`;
   }
-switch (holiday) {
-  case "Shavuos":
-    holiday = "Rus";
-    break;
-  case "Pesach":
-    holiday = "Haggadah";
-    break;
-  case "Purim":
-     holiday = "Megillah";
-    break;
-  default:
-    holiday = "";
-}
-if (holiday != "") {
-  parsha.push(holiday);
-  if (subject != "") {
-    subject = `${subject} & ${holiday}`;
+  if (holiday != "") {
+    parsha.push(holiday);
+    if (subject != "") {
+      subject = `${subject} & ${holiday}`;
+    }
+    else {
+      subject = `${holiday} Questions`;
+    }
   }
-  else {
-    subject = `${holiday} Questions`;
-  }
-}
   var folders = DriveApp.getFolderById("1ihDno5us7sLY6NH5fmqORPuplD8Uz_66");
   var files = folders.getFiles();
   var file;
   while (files.hasNext()) {
     file = files.next();
     for (var name in parsha) {
-        if (file.getName().includes(parsha[name])) {
+      if (file.getName().includes(parsha[name])) {
         attachments.push(file);
       }
     }
@@ -102,12 +122,15 @@ if (holiday != "") {
       break;
     }
   }
+  console.log("Subject: " + subject);
   for (var file in attachments) {
     console.log(attachments[file].getName());
   }
-  GmailApp.sendEmail("[EMAIL REMOVED]", subject, "",{
+  GmailApp.sendEmail([EMAIL REMOVED], subject, "",{
         from: "aribennett1@gmail.com",
         attachments: attachments,
         name: "Ari Bennett"
       });
+
+  console.log(`RemainingDailyQuota: ${MailApp.getRemainingDailyQuota()}`);
 }
